refactor: types index file
Modify LanguageType structure to add HERO_TAGLINE, HERO_CTA.

And also modify the HEADER structure and convert it to a type and use
it.

Define into the structure also the SECTIONS that uses SectionItem type,
and SKILLS the SkillSectionItem.

In the Skill type make icon optional, and add the type prop with is own
type called SkillType that is a type with defined values.

Modify Job type and remove some props (rename them)

Also remove ProjectProps Type

To finish, add and export SkillType type